backend/database/pyproject.toml
[project]
name = "alex-database"
version = "0.1.0"
description = "Database package for Alex Financial Planner with Aurora Data API"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.40.8",
    "google-cloud-secret-manager>=2.20.0",
    "psycopg>=3.3.3",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
